Upgrade to Swift 2/Xcode 7
Mostly left this to the automatic upgrade mechanism. Getting a couple of warnings which need to be tracked down, but the app runs as expected.

One major snag is the new requirement to use HTTPS when making requests. I've temporarily set the NSAllowsArbitraryLoads flag to get around this, but this is presumably not allowed in apps submitted to the App Store, which means I need to update the server to cache the images and serve them over HTTPS.